feat(etherscan_verify): Refacto
