+++ typescript/pragma-deployer/src/utils/index.ts
@@ -1,0 +1,1 @@
+export { ETHERSCAN_VERIFIER_URL, verifyContract } from "./evm_verify";
